Updates to Choose Crew and Work A day Methods
