+++ LDAR_Sim/inputs_template/P_dev_OGI.txt
@@ -31,7 +31,7 @@ P_dev_OGI = {
     },
     'program_name': 'P_dev_OGI',
     'n_simulations': 2, 	# Minimum of 2 simulations to get batch plots
-    'timesteps': 1000,  	# Up to ~5600 for 16 year nc file
+    'timesteps': 2000,  	# Up to ~5600 for 16 year nc file
     'n_processes': None,
     'start_year': 2017,
     'UTC_offset': -6,
