+++ LDAR_Sim/inputs_template/P_dev_aircraft.txt
@@ -60,7 +60,7 @@ P_dev_aircraft = {
     },
     'program_name': 'P_dev_aircraft',
     'n_simulations': 2, 	# Minimum of 2 simulations to get batch plots
-    'timesteps': 1000,  	# Up to ~5600 for 16 year nc file
+    'timesteps': 2000,  	# Up to ~5600 for 16 year nc file
     'n_processes': None,
     'start_year': 2017,
     'UTC_offset': -6,
