+++ LDAR_Sim/model_code/dev_OGI_company.py
@@ -1,7 +1,7 @@
 
 
-from methods.base_company import BaseCompany
-from methods.mobile_crew import MobileCrew
+from methods.company import BaseCompany
+from methods.crew import BaseCrew
 
 
 class dev_OGI_company(BaseCompany):
@@ -13,16 +13,12 @@ class dev_OGI_company(BaseCompany):
         super(dev_OGI_company, self).__init__(state, parameters, config, timeseries)
         # --- Custom Init ---
         # -------------------
-        # Initiate Crews - This is done outside of base class
-        for i in range(config['n_crews']):
-            self.crews.append(Crew(state, parameters, config,
-                                   timeseries, self.deployment_days, id=i + 1))
 
     # --- Custom Methods ---
     # ----------------------
 
 
-class Crew(MobileCrew):
+class Crew(BaseCrew):
     """ Test Company module. Initialize each crew for company.
         Inherits base method base class company
