+++ LDAR_Sim/model_code/ldar_sim.py
@@ -33,6 +33,25 @@ from daylight_calculator import DaylightCalculatorAve
 from generic_functions import make_maps
 from utils.distributions import leak_rvs, fit_dist
 
+# --- Performance Test ---
+import time
+from functools import wraps
+
+
+def timer(method):
+    @wraps(method)
+    def timed(self, *args, **kwargs):
+        ts = time.time()
+        result = method(self, *args, **kwargs)
+        te = time.time()
+        try:
+            self.state['perform_test'][method.__name__] += te - ts
+        except KeyError:
+            self.state['perform_test'].update({method.__name__: te - ts})
+        return result
+    return timed
+# --- Performance Test ---
+
 
 class LdarSim:
     def __init__(self, state, params, timeseries):
@@ -44,7 +63,21 @@ class LdarSim:
         self.parameters = params
         self.timeseries = timeseries
         self.active_leaks = []
+        #  --- timeseries variables ---
+        timeseries['total_daily_cost'] = np.zeros(params['timesteps'])
+        timeseries['repair_cost'] = np.zeros(params['timesteps'])
+        timeseries['verification_cost'] = np.zeros(params['timesteps'])
+        timeseries['operator_redund_tags'] = np.zeros(self.parameters['timesteps'])
+        timeseries['operator_tags'] = np.zeros(self.parameters['timesteps'])
 
+        # Configure sensitivity analysis, if requested (code block must remain here -
+        # after site initialization and before method initialization)
+        if params['sensitivity']['perform']:
+            self.sensitivity = Sensitivity(params, timeseries, state)
+
+        #  --- state variables ---
+        state['perform_test'] = {}
+        state['candidate_flags'] = {}
         # Read in data files
         state['empirical_counts'] = np.array(pd.read_csv(
             params['working_directory'] + params['count_file']).iloc[:, 0])
@@ -130,18 +163,6 @@ class LdarSim:
                 site['subtype_code'] = 0
             site['leak_rate_dist'] = params['dists'][int(site['subtype_code'])]['dist']
             site['leak_rate_units'] = params['dists'][int(site['subtype_code'])]['units']
-
-        # Additional timeseries variables
-        timeseries['total_daily_cost'] = np.zeros(params['timesteps'])
-        timeseries['repair_cost'] = np.zeros(params['timesteps'])
-        timeseries['verification_cost'] = np.zeros(params['timesteps'])
-        timeseries['operator_redund_tags'] = np.zeros(self.parameters['timesteps'])
-        timeseries['operator_tags'] = np.zeros(self.parameters['timesteps'])
-
-        # Configure sensitivity analysis, if requested (code block must remain here -
-        # after site initialization and before method initialization)
-        if params['sensitivity']['perform']:
-            self.sensitivity = Sensitivity(params, timeseries, state)
 
         # Initialize method(s) to be used; append to state
         for m_key, m_obj in params['methods'].items():
@@ -242,7 +263,7 @@ class LdarSim:
 
             # Change negatives to zero
             state['empirical_vents'] = [0 if i < 0 else i for i in state['empirical_vents']]
-
+        state['perform_test']['init_time'] = time.time()
         return
 
     def update(self):
@@ -258,6 +279,7 @@ class LdarSim:
         self.report()  # Assemble any reporting about model state
         return
 
+    @timer
     def update_state(self):
         """
         update the state of active leaks
@@ -286,6 +308,7 @@ class LdarSim:
         self.timeseries['active_leaks'].append(len(self.active_leaks))
         self.timeseries['datetime'].append(self.state['t'].current_date)
 
+    @timer
     def add_leaks(self):
         """
         add new leaks to the leak pool
@@ -333,6 +356,7 @@ class LdarSim:
 
         return
 
+    @timer
     def deploy_crews(self):
         """
         Loop over all your methods in the simulation and ask them to find some leaks.
@@ -347,6 +371,7 @@ class LdarSim:
 
         return
 
+    @timer
     def repair_leaks(self):
         """
         Repair tagged leaks and remove from tag pool.
@@ -383,6 +408,7 @@ class LdarSim:
 
         return
 
+    @timer
     def report(self):
         """
         Daily reporting of leaks, repairs, and emissions.
@@ -403,6 +429,7 @@ class LdarSim:
 
         return
 
+    @timer
     def finalize(self):
         """
         Compile and write output files.
