+++ LDAR_Sim/model_code/methods/sensors/default.py
@@ -5,19 +5,16 @@ from method_functions import measured_rate
 def detect_emissions(self, site, leaks_present, equipment_rates, site_true_rate, venting):
     m_name = self.config['label']
     site_measured_rate = 0
-    is_leak_detected = False
 
     if self.config["measurement_scale"] == "site":
         if site_true_rate > (self.config['MDL']):
             # If source is above follow-up threshold, calculate measured rate using QE
             site_measured_rate = measured_rate(site_true_rate, self.config['QE'])
-            is_leak_detected = True
     elif self.config["measurement_scale"] == "equipment":
         for rate in equipment_rates:
             if rate > (self.config['MDL']):
                 equip_measured_rate = measured_rate(rate, self.config['QE'])
                 site_measured_rate += equip_measured_rate
-                is_leak_detected = True
     elif self.config["measurement_scale"] == "leak":
         # If measurement scale is a leak, all leaks will be tagged
         for leak in leaks_present:
@@ -33,10 +30,8 @@ def detect_emissions(self, site, leaks_present, equipment_rates, site_true_rate,
                     leak['tagged_by_crew'] = self.crewstate['id']
                     self.state['tags'].append(leak)
                     site_measured_rate += measured_rate(leak['rate'], self.config['QE'])
-                    is_leak_detected = True
             else:
                 site[self.config['label'] + '_missed_leaks'] += 1
-        return is_leak_detected
 
     # If source is above follow-up threshold
     if site_measured_rate > self.config['follow_up_thresh']:
@@ -48,9 +43,10 @@ def detect_emissions(self, site, leaks_present, equipment_rates, site_true_rate,
             'site_measured_rate': site_measured_rate,
             'venting': venting
         }
-        self.candidate_flags.append(site_dict)
+        # self.candidate_flags.append(site_dict)
 
     else:
+        site_dict = None
         site['{}_missed_leaks'.format(m_name)] += len(leaks_present)
 
-    return is_leak_detected
+    return site_dict
