+++ LDAR_Sim/model_code/utils/performance.py
@@ -1,0 +1,16 @@
+import time
+from functools import wraps
+
+
+def timer(method):
+    @wraps(method)
+    def timed(self, *args, **kwargs):
+        ts = time.time()
+        result = method(self, *args, **kwargs)
+        te = time.time()
+        try:
+            self.state['perform_test'][method.__name__] += te - ts
+        except KeyError:
+            self.state['perform_test'].update({method.__name__: te - ts})
+        return result
+    return timed
